backend/requirements.txt
aiofiles==25.1.0
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
async-timeout==5.0.1
attrs==25.4.0
bcrypt==4.1.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
cssselect2==0.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.110.1
fonttools==4.61.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
jsonschema==4.26.0
MarkupSafe==3.0.3
motor==3.3.1
multidict==6.7.1
openpyxl==3.1.5
packaging==26.0
passlib==1.7.4
pillow==12.1.1
propcache==0.4.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.12.1
PyJWT==2.11.0
pymongo==4.5.0
pyparsing==3.3.2
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
redis==7.2.0
referencing==0.37.0
reportlab==4.4.10
requests==2.32.5
rpds-py==0.30.0
sentry-sdk==2.53.0
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
tenacity==9.1.4
tinycss2==1.5.1
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.25.0
watchfiles==1.1.1
weasyprint==68.1
webencodings==0.5.1
websockets==15.0.1
xlsxwriter==3.2.9
yarl==1.22.0
zopfli==0.4.1
